moved user model into models directory
